refactor: unified prepare and build-schematics executors
