+++ libs/ui5-angular-theming/package.json
@@ -1,6 +1,13 @@
 {
   "name": "@ui5/theming-ngx",
   "version": "0.0.1",
+  "license": "Apache-2.0",
+  "author": "SAP SE (https://www.sap.com)",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/SAP/ui5-webcomponents-ngx.git",
+    "directory": "libs/ui5-angular-theming"
+  },
   "peerDependencies": {
     "@angular/common": "^15.0.0",
     "@angular/core": "^15.0.0"
