+++ libs/ui5-angular-theming/project.json
@@ -5,6 +5,9 @@
   "sourceRoot": "libs/ui5-angular-theming/src",
   "prefix": "ui-angular",
   "targets": {
+    "prepare": {
+      "executor": "@ui5/webcomponents-nx:prepare"
+    },
     "sync": {
       "executor": "@ui5/webcomponents-nx:sync",
       "options": {
@@ -14,7 +17,7 @@
     "build": {
       "executor": "@nrwl/angular:package",
       "outputs": [
-        "{workspaceRoot}/dist/{projectRoot}"
+        "dist/libs/ui5-angular-theming"
       ],
       "options": {
         "project": "libs/ui5-angular-theming/ng-package.json"
