+++ libs/webcomponents-nx/executors.json
@@ -6,11 +6,6 @@
       "schema": "./src/executors/sync/schema.json",
       "description": "Sync executor"
     },
-    "build-schematics": {
-      "implementation": "./src/executors/build-schematics/executor",
-      "schema": "./src/executors/build-schematics/schema.json",
-      "description": "Builds the schematics for the library"
-    },
     "prepare": {
       "implementation": "./src/executors/prepare/executor",
       "schema": "./src/executors/prepare/schema.json",
