--- libs/webcomponents-nx/src/executors/build-schematics/executor.spec.ts
@@ -1,11 +1,0 @@
-import { BuildSchematicsExecutorSchema } from './schema';
-import executor from './executor';
-
-const options: BuildSchematicsExecutorSchema = {};
-
-describe('BuildSchematics Executor', () => {
-  it('can run', async () => {
-    const output = await executor(options);
-    expect(output.success).toBe(true);
-  });
-});
