--- libs/webcomponents-nx/src/executors/build-schematics/schema.d.ts
@@ -1,1 +1,0 @@
-export interface BuildSchematicsExecutorSchema {} // eslint-disable-line
