--- libs/webcomponents-nx/src/executors/build-schematics/schema.json
@@ -1,10 +1,0 @@
-{
-  "$schema": "http://json-schema.org/schema",
-  "version": 2,
-  "cli": "nx",
-  "title": "BuildSchematics executor",
-  "description": "",
-  "type": "object",
-  "properties": {},
-  "required": []
-}
