+++ libs/webcomponents-nx/src/executors/prepare/prepareOptions.ts
@@ -1,0 +1,10 @@
+import {SchematicsOptions} from "./schematicsOptions";
+
+export interface PrepareOptions {
+  schematics?: SchematicsOptions;
+  packageJsonPath?: string;
+  versionsOverrides?: {
+    projectVersion?: string;
+  };
+  distPath: string;
+}
