+++ libs/webcomponents-nx/src/executors/prepare/schematicsOptions.ts
@@ -1,0 +1,4 @@
+export interface SchematicsOptions {
+  collection: string;
+  tsConfig: string;
+}
